+++ Services/Implementations/UserService.cs
@@ -1,31 +1,41 @@
 ﻿using SkillStackCSharp.Services.Interfaces;
-using SkillStackCSharp.Models;
 using SkillStackCSharp.Repositories.Interfaces;
-using SkillStackCSharp.DTOs;
-using System.Collections.Generic;
+using SkillStackCSharp.Models;
+using SkillStackCSharp.DTOs.UserDTOs;
+using Microsoft.AspNetCore.Identity;
+using SkillStackCSharp.Constants;
 
 namespace SkillStackCSharp.Services.Implementations
 {
     public class UserService: IUserService
     {
         private readonly IUserRepository _userRepository;
+        private readonly UserManager<User> _userManager;
+        private readonly RoleManager<IdentityRole> _roleManager;
 
-        public UserService(IUserRepository userRepository)
+        public UserService(IUserRepository userRepository, UserManager<User> userManager, RoleManager<IdentityRole> roleManager)
         {
             _userRepository = userRepository;
+            _userManager = userManager;
+            _roleManager = roleManager;
         }
 
         public async Task<IEnumerable<UserDTO>> GetAllUsersAsync()
         {
             var users = await _userRepository.GetAllUsersAsync();
-            foreach (var user in users)
-            {
-                
-            }
-            //var userDTO = new UserDTO() { 
 
-            //};
-            return null;
+            // Map User entities to UserDTO
+            var userDto = users.Select(user => new UserDTO
+            {
+                Id = user.Id,
+                UserName = user.UserName,
+                Email = user.Email,
+                FirstName = user.FirstName,
+                LastName =  user.LastName,
+                EmailConfirmed = user.EmailConfirmed
+            });
+
+            return userDto;
         }
 
         public async Task<UserDTO> GetUserByIdAsync(string id)
@@ -39,30 +49,93 @@ namespace SkillStackCSharp.Services.Implementations
                 Email = user.Email,
                 FirstName = user.FirstName,
                 LastName = user.LastName,
-                Username = user.UserName,
+                UserName = user.UserName,
                 EmailConfirmed = user.EmailConfirmed,
             };
 
             return userDTO;
         }
 
-        public async Task CreateUserAsync(UserDTO user)
+        public async Task<UserDTO> CreateUserAsync(CreateUserDTO createUserDTO)
         {
-            //_userRepository.AddUser(user);
-            //await _userRepository.SaveChangesAsync(); // Save changes to the database
+            var user = new User()
+            {
+                UserName = createUserDTO.UserName,
+                Email = createUserDTO.Email,
+                FirstName = createUserDTO.FirstName,
+                LastName = createUserDTO.LastName
+            };
+
+            var result = await _userManager.CreateAsync(user, createUserDTO.Password);
+
+            if (!result.Succeeded)
+                throw new Exception("User creation failed: " + string.Join(", ", result.Errors.Select(e => e.Description)));
+
+            if(createUserDTO.Role == null)
+                await _userManager.AddToRoleAsync(user, UserRoles.User);
+            else if (await _roleManager.RoleExistsAsync(createUserDTO.Role))
+                await _userManager.AddToRoleAsync(user, createUserDTO.Role);
+        
+            await _userRepository.SaveChangesAsync();
+
+            var userDTO = new UserDTO() { 
+                Id=user.Id,
+                Email = user.Email,
+                FirstName = user.FirstName,
+                LastName = user.LastName,
+                UserName = user.UserName,
+                EmailConfirmed = user.EmailConfirmed,
+            };
+
+            return userDTO;
         }
 
-        public async Task UpdateUserAsync(UserDTO user)
+        public async Task<UserDTO> UpdateUserAsync(string id, UpdateUserDTO updateUserDTO)
         {
-            //_userRepository.UpdateUser(user);
-            //await _userRepository.SaveChangesAsync(); // Save changes to the database
+            var user = await _userRepository.GetUserByIdAsync(id);
+
+            if (user == null)
+                return null;
+
+            // Update the user properties
+            user.FirstName = updateUserDTO.FirstName ?? user.FirstName;
+            user.LastName = updateUserDTO.LastName ?? user.LastName;
+            user.UserName = updateUserDTO.UserName ?? user.UserName;
+            user.Email = updateUserDTO.Email ?? user.Email;
+
+            _userRepository.UpdateUser(user);
+            await _userRepository.SaveChangesAsync();
+
+            var userDTO = new UserDTO() { 
+                Id = user.Id,
+                UserName = user.UserName,
+                Email = user.Email,
+                FirstName = user.FirstName,
+                LastName = user.LastName,
+                EmailConfirmed = user.EmailConfirmed,
+            };
+
+            return userDTO;
         }
 
-        public async Task DeleteUserAsync(UserDTO userDTO)
+        public async Task<bool> DeleteUserAsync(string id)
         {
-            var user = await _userRepository.GetUserByIdAsync(userDTO.Id);
+            var user = await _userRepository.GetUserByIdAsync(id);
+
+            if (user == null)
+                return false;
+
             _userRepository.RemoveUser(user);
             await _userRepository.SaveChangesAsync(); // Save changes to the database
+
+            return true;
+        }
+
+        public async Task<IdentityResult> ChangePasswordAsync(string id,string oldPassword, string newPassword)
+        {
+            var user = await _userManager.FindByIdAsync(id);
+            var result = await _userManager.ChangePasswordAsync(user, oldPassword, newPassword);
+            return result;
         }
     }
 }
